pyproject.toml
[tool.poetry]
name = "glitchtip-backend"
version = "0.1.0"
description = "Django backend that powers GlitchTip, an open source reimplementation of Sentry"
authors = ["David Burke"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "~5.0.1"
djangorestframework = "^3.12.1"
django-allauth = "~0.61.0"
django-cors-headers = "^4.0.0"
django-environ = "~0.11.2"
django-extensions = "^3.0.2"
django-organizations = "^2.3.1"
django-storages = {extras = ["boto3", "azure", "google"], version = "^1.12.3"}
drf-nested-routers = "^0.93.3"
uWSGI = "^2.0.18"
whitenoise = "^6.0.0"
sentry-sdk = "^1.0.0"
django-filter = "^22.1"
celery = {version = "~5.3.0", extras = ["redis"]}
django-csp = "^3.6"
dj-stripe = "~2.8.0"
django-anymail = "^10.2"
dj-rest-auth = "~5.0.0"
user-agents = "^2.1"
django-ipware = "^6.0.0"
anonymizeip = "^1.0.0"
symbolic = "^10.0.0"
django-rest-mfa = "^1.2.2"
aiohttp = "^3.7.4"
google-cloud-logging = "^3.0.0"
django-prometheus = "^2.2.0"
django-sql-utils = "^0.7.0"
django-redis = {extras = ["hiredis"], version = "^5.2.0"}

uwsgi-chunked = "^0.1.8"
django-import-export = "^3.0.0"
psycopg = {extras = ["c"], version = "^3.1.12"}
uvicorn = "^0.29.0"
gunicorn = "^21.2.0"
boto3 = "1.34.69"
django-ninja = "^1.0.1"
orjson = "^3.9.9"
celery-batches = "^0.8.1"

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.2"
locust = "^2.10.1"
freezegun = "^1.1.0"
rope = "^1.3.0"
model_bakery = "^1.1.1"
django-debug-toolbar = "^4.0.0"
django-stubs = {extras = ["compatible-mypy"], version = "^4.0.0"}
django-sslserver = "^0.22"
aioresponses = "^0.7.6"
requests-mock = "^1.9.3"
memray = "^1.9.1"
ruff = "^0.3.0"
# Needed for test --parallel
tblib = "^3.0.0"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.per-file-ignores]
"*/migrations/*.py" = ["I"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "glitchtip.settings"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
